+++ docs-gen/content/docs/examples/_index.md
@@ -1,0 +1,8 @@
+---
+title: "Examples"
+weight: 2
+---
+
+All code samples can be found on [GitHub]({{< param "github_repo" >}}/blob/{{< param "github_branch" >}}/{{< param "github_subdir" >}}/includes/quickstart).
+
+We assume that you are already familiar with TFX and Kubeflow and that you have access to a running instance of Kubeflow Pipelines.
