+++ docs-gen/content/docs/reference/_index.md
@@ -1,4 +1,4 @@
 ---
 title: "Reference"
-weight: 2
+weight: 3
 ---
