+++ docs-gen/includes/dependent/penguin_training/Dockerfile
@@ -1,0 +1,7 @@
+# TFX build
+FROM tensorflow/tfx:1.2.0
+
+WORKDIR /pipeline
+COPY ./*.py ./
+
+ENV PYTHONPATH="/pipeline:${PYTHONPATH}"
