+++ docs-gen/includes/dependent/resources/runconfiguration_training.yaml
@@ -1,0 +1,16 @@
+apiVersion: pipelines.kubeflow.org/v1alpha5
+kind: RunConfiguration
+metadata:
+    name: penguin-pipeline-ensembled-run
+spec:
+    run:
+      pipeline: penguin-pipeline-ensembled
+      runtimeParameters:
+      - name: examples_location
+        valueFrom:
+          runConfigurationRef:
+            name: penguin-pipeline-examples
+            outputArtifact: examples
+    triggers:
+      runConfigurations:
+        - penguin-pipeline-examples
