+++ docs/sitemap.xml
@@ -1,1 +1,1 @@
-<?xml version="1.0" encoding="utf-8" standalone="yes"?><urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml"><url><loc>https://sky-uk.github.io/kfp-operator/docs/reference/configuration/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/docs/getting-started/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/docs/getting-started/overview/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/docs/reference/resources/pipeline/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/docs/getting-started/installation/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/docs/reference/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/docs/reference/resources/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/docs/reference/resources/runconfiguration/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/docs/getting-started/example/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/docs/reference/resources/run/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/docs/reference/run-completion/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/docs/reference/resources/experiment/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/docs/reference/providers/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/docs/reference/debugging/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/categories/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/docs/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/tags/</loc></url></urlset>
+<?xml version="1.0" encoding="utf-8" standalone="yes"?><urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml"><url><loc>https://sky-uk.github.io/kfp-operator/docs/reference/configuration/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/docs/getting-started/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/docs/getting-started/overview/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/docs/reference/resources/pipeline/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/docs/examples/pipeline_training/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/docs/examples/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/docs/getting-started/installation/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/docs/examples/pipeline_dependencies/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/docs/reference/resources/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/docs/reference/resources/runconfiguration/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/docs/reference/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/docs/reference/resources/run/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/docs/reference/run-completion/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/docs/reference/resources/experiment/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/docs/reference/providers/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/docs/reference/debugging/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/categories/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/docs/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/</loc></url><url><loc>https://sky-uk.github.io/kfp-operator/tags/</loc></url></urlset>
